top python projects on GitHub
This project can show you projects with the most quantyti of stars, which has writed on Python language. And you can see it on real time because i used API requests. If you are interested some project, so you can click on the project name which is also a link to project.